Fix in new metaSQL parser code for a double-to-int conversion warning.
git-svn-id: http://svn.code.sf.net/p/postbooks/code/openrpt/trunk@12824 ea1a15ef-c238-0410-93cb-fa28e9f07733

+++ MetaSQL/metasqlqueryparser.cpp
@@ -123,7 +123,7 @@ class MetaSQLFunction : public MetaSQLOutput {
                         _valid = 1;
                         _noOutput = 1;
                         if(params.size() >= 1)
-                            _nBreaks = convertToDouble(params[0]);
+                            _nBreaks = (int) convertToDouble(params[0]);
                         if(_nBreaks < 1) _nBreaks = 1;
                         break;
                     default:
